pyproject.toml
[project]
name = "gemini-search"
version = "0.1.0"
description = "CLI RAG chat app using Gemini with Google Drive files"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "click>=8.1.0",
    "prompt_toolkit>=3.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "itsdangerous>=2.1.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
gemini-search = "gemini_search.cli:main"
gemini-search-web = "gemini_search.web:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"